Updated diagnostic notebook showing colors of cluster environment. Current status of branch is as follows.
I've taken a reasonable look at the remapping functions for both centrals and satellites. The main thing remaining to do on that end is to fix up the internals of the remap_cluster_bcg_gr_ri_color function so that the restframe colors of cluster BCGs blue-shift in the same way as the all-galaxy red sequence. Currently, cluster satellites only have an elevated red fraction, but not a reddened peak relative to the all-galaxy red sequence. Probably it would be a good idea to do something exactly analogous for the cluster BCGs, but redshifting their colors by ~0.05dex relative to the peak of the all-galaxy red sequence.

One thing that continue to surprise me is that remapping the satellite colors seems to have virtually zero impact whatsoever on red/blue clustering. If I don't remap at all, or if I remap with 100% probability, I seem to get statistically consistent clustering in the v4p4_color_calibration_snapshots.ipynb notebook. That seems like it can't be right. It would be good to know what's going on there before merging this branch. For example, check whether the satellite fraction actually gets boosted for red samples when remapping all satellites with 100% probability.

Before this branch is merged, make sure to clean up the now-obsolete analytical_colors.py module, and all references to it throughout the package.

Once that's settled, the next step will be to do a trial run with Dan's matchup code, to verify that my shifted restframe colors are sufficient to shift the observed color distributions to match DEEP2. That can be done with just the z=0.9 snap. If it checks out, then give the mock one last jupyter-run-all-validation, and handoff to Dan to make v4.4.